+++ skbio/diversity/tests/test_driver.py
@@ -280,32 +280,29 @@ class BetaDiversityTests(TestCase):
 
     def test_invalid_input(self):
         # number of ids doesn't match the number of samples
-        error_msg = ("Number of rows in ``counts`` must be equal to number of"
-                     " provided ``ids``.")
+        error_msg = ("Number of rows")
         with six.assertRaisesRegex(self, ValueError, error_msg):
             beta_diversity(self.table1, list('AB'), 'euclidean')
 
         # unknown metric provided
-        error_msg = "Unknown Distance Metric: not-a-metric"
+        error_msg = "not-a-metric"
         with six.assertRaisesRegex(self, ValueError, error_msg):
             beta_diversity('not-a-metric', self.table1)
 
         # 3-D list provided as input
-        error_msg = ("Only 1-D and 2-D array-like objects can be provided as "
-                     "input. Provided object has 3 dimensions.")
+        error_msg = ("Only 1-D and 2-D")
         with six.assertRaisesRegex(self, ValueError, error_msg):
             beta_diversity('euclidean', [[[43]]])
 
         # negative counts
-        error_msg = "Counts vector cannot contain negative values."
+        error_msg = "negative values."
         with six.assertRaisesRegex(self, ValueError, error_msg):
             beta_diversity('euclidean', [[0, 1, 3, 4], [0, 3, -12, 42]])
         with six.assertRaisesRegex(self, ValueError, error_msg):
             beta_diversity('euclidean', [[0, 1, 3, -4], [0, 3, 12, 42]])
 
         # additional kwargs
-        error_msg = ("pdist\(\) got an unexpected keyword argument "
-                     "'not_a_real_kwarg'")
+        error_msg = ("'not_a_real_kwarg'")
         with six.assertRaisesRegex(self, TypeError, error_msg):
             beta_diversity('euclidean', [[0, 1, 3], [0, 3, 12]],
                            not_a_real_kwarg=42.0)
@@ -317,8 +314,6 @@ class BetaDiversityTests(TestCase):
             beta_diversity('weighted_unifrac', [[0, 1, 3], [0, 3, 12]],
                            not_a_real_kwarg=42.0, tree=self.tree1,
                            otu_ids=['O1', 'O2', 'O3'])
-        error_msg = ("weighted_unifrac\(\) got an unexpected keyword argument "
-                     "'not_a_real_kwarg'")
         with six.assertRaisesRegex(self, TypeError, error_msg):
             beta_diversity(weighted_unifrac, [[0, 1, 3], [0, 3, 12]],
                            not_a_real_kwarg=42.0, tree=self.tree1,
@@ -581,30 +576,30 @@ class BetaDiversityTests(TestCase):
                 if id1 != id2:
                     self.assertNotEqual(dm1[id1, id2], dm2[id1, id2])
 
-    def test_alt_pdist_func(self):
-        # confirm that pdist_func is actually being used
+    def test_alt_pairwise_func(self):
+        # confirm that pairwise_func is actually being used
         def not_a_real_pdist(counts, metric):
             return [[0.0, 42.0], [42.0, 0.0]]
         dm1 = beta_diversity('unweighted_unifrac', self.table1,
                              otu_ids=self.oids1, tree=self.tree1,
-                             pdist_func=not_a_real_pdist)
+                             pairwise_func=not_a_real_pdist)
         expected = DistanceMatrix([[0.0, 42.0], [42.0, 0.0]])
         self.assertEqual(dm1, expected)
 
         dm1 = beta_diversity('weighted_unifrac', self.table1,
                              otu_ids=self.oids1, tree=self.tree1,
-                             pdist_func=not_a_real_pdist)
+                             pairwise_func=not_a_real_pdist)
         expected = DistanceMatrix([[0.0, 42.0], [42.0, 0.0]])
         self.assertEqual(dm1, expected)
 
         dm1 = beta_diversity(unweighted_unifrac, self.table1,
                              otu_ids=self.oids1, tree=self.tree1,
-                             pdist_func=not_a_real_pdist)
+                             pairwise_func=not_a_real_pdist)
         expected = DistanceMatrix([[0.0, 42.0], [42.0, 0.0]])
         self.assertEqual(dm1, expected)
 
         dm1 = beta_diversity("euclidean", self.table1,
-                             pdist_func=not_a_real_pdist)
+                             pairwise_func=not_a_real_pdist)
         expected = DistanceMatrix([[0.0, 42.0], [42.0, 0.0]])
         self.assertEqual(dm1, expected)